+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added `--upstream` (dependencies) and `--downstream` (dependents) options to `moon ci` to control
+  the graph depth for affected tracking.
+
 #### 🐞 Fixes
 
 - Fixed an issue where a task `mutex` would not be respected, and tasks would still run in parallel.